requirements/base.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mysql-connector-python>=8.2.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
